fix: hover animation
