+++ src/components/EnterpriseApplyCenter.jsx
@@ -1,4 +1,3 @@
-import './css/EnterpriseApplyCenter.css';
 import './css/SingInCenter.css';
 import React, { useState, useEffect, useRef, useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
