+++ src/pages/Apply/EnterpriseApply.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Header from '../component/Header';
-import EnterpriseApplyCenter from '../components/EnterpriseApplyCenter';
+import Header from '../../component/Header';
+import EnterpriseApplyCenter from '../../components/EnterpriseApplyCenter';
 const EnterpriseApply = () => {
   return (
     <>
