+++ src/pages/Apply/JobNoticeApply.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import JobNoticeApplyCenter from '../../components/JobNoticeApplyCenter';
+import Header from '../../component/Header';
+import Footer from '../../component/Footer';
+const JobNoticeApply = () => {
+  return (
+    <div>
+      <JobNoticeApplyCenter />
+    </div>
+  );
+};
+
+export default JobNoticeApply;
