+++ src/pages/Enterprises/CompanyDetails.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from '../component/Header';
-import Footer from '../component/Footer';
-import CompanyDetailsCenter from '../components/CompanyDetailsCenter';
+import Header from '../../component/Header';
+import Footer from '../../component/Footer';
+import CompanyDetailsCenter from './CompanyDetailsCenter';
 const CompanyDetails = () => {
   return (
     <div>
