--- src/pages/JobNotice.jsx
@@ -1,14 +1,0 @@
-import CompanyList from './Enterprises/components/EnterprisesList';
-import Header from '../component/Header';
-import JobNoticeList from '../components/JobNoticeList';
-import RecentReview from '../components/RecentReviewList';
-import Footer from '../component/Footer';
-function JobNotice() {
-  return (
-    <div>
-      <JobNoticeList />
-    </div>
-  );
-}
-
-export default JobNotice;
