+++ src/pages/Jobs/JobNoticeDetails.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from '../component/Header';
-import JobNoticeDetailsCenter from '../components/JobNoticeDetailsCenter';
-import Footer from '../component/Footer';
+import Header from '../../component/Header';
+import JobNoticeDetailsCenter from '../../components/JobNoticeDetailsCenter';
+import Footer from '../../component/Footer';
 
 const JobNoticeDetails = () => {
   return (
