+++ src/pages/LegalChat/IegalAdvice.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Header from '../../component/Header';
+import Footer from '../../component/Footer';
+import IegalAdviceCenter from '../../components/IegalAdviceCenter';
+const IegalAdvice = () => {
+  return (
+    <div>
+      <IegalAdviceCenter />
+    </div>
+  );
+};
+
+export default IegalAdvice;
